refactor adCard, ad home page (#2057)
