+++ libs/gi/page-home/src/index.tsx
@@ -1,4 +1,5 @@
 import { CardThemed } from '@genshin-optimizer/common/ui'
+import { AdCard } from '@genshin-optimizer/gi/ui'
 import DescriptionIcon from '@mui/icons-material/Description'
 import {
   Box,
@@ -39,6 +40,7 @@ export default function PageHome() {
         >
           <QuickLinksCard />
           <ResinCard />
+          <AdCard dataAdSlot="6687816711" />
         </Grid>
         <Grid
           item
