+++ libs/gi/ui/src/components/ad/GoAd/GOAd.tsx
@@ -1,7 +1,7 @@
 import { Typography } from '@mui/material'
 import { Box } from '@mui/system'
 import type { ReactNode } from 'react'
-import { toMSite } from './util'
+import { toMainSite } from '../util'
 
 export function GOAd({ children }: { children: ReactNode }) {
   return (
@@ -17,7 +17,7 @@ export function GOAd({ children }: { children: ReactNode }) {
         minHeight: '100%',
         minWidth: '100%',
       }}
-      onClick={toMSite}
+      onClick={toMainSite}
     >
       {children}
       <Typography color="black">
