+++ libs/gi/ui/src/components/ad/index.ts
@@ -1,1 +1,2 @@
+export * from './AdCard'
 export * from './AdWrapper'
